requirements.txt
dify_plugin
azure-ai-vision-imageanalysis==1.0.0
